unit tests for gates
