+++ tests/qasm3_qir/conftest.py
@@ -15,4 +15,4 @@ without needing to import them (pytest will automatically discover them).
 """
 # pylint: disable=unused-import,wildcard-import,unused-wildcard-import
 
-from .fixtures.basic_gates import *
+from .fixtures.gates import *
