+++ tests/qasm3_qir/fixtures/resources/custom_gate_simple.qasm
@@ -1,0 +1,12 @@
+OPENQASM 3;
+include "stdgates.inc";
+
+gate custom(a) p, q {
+    h p;
+    z q;
+    rx(a) p;
+    cx p,q;
+}
+
+qubit[2] q;
+custom(0.1) q[0], q[1];
